ISLAND FINDING: Refactor extract shared parent class
There might be architecturally beter approaches to reduce
duplication, but for now this shared inheritance allows for reduction
and re-use of a lot of shared helper methods.

The processGrid method is the method in which the traversal of the grid
and the visiting of specific cells should take place. This method is made
to follow a template method pattern to not couple the extending classes
based on that which they must provide/solve, otherwise expected ripple effects.